# proyecto-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Muestra un listado de comentarios",
                "description": "Devuelve un listado de comentarios",
                "operationId": "getCommentsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "añade un comentario",
                "description": "devuelve los datos de un comentario",
                "operationId": "addComment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario no autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/comments/{comment}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "obtiene un comentario",
                "description": "devuelve un comentario",
                "operationId": "getComment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "id del comentario a mostrar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Edita un comentario",
                "description": "devuelve los datos de un comentario",
                "operationId": "updateComment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "Identificador del recurso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario no autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    },
                    {
                        "sanctum": []
                    },
                    {
                        "commentOwnerOrAdmin": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Borra un comentario",
                "description": "Borra un comentario y devuelve un 204",
                "operationId": "deleteComment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "Identificador del recurso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario no autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    },
                    {
                        "sanctum": []
                    },
                    {
                        "commentOwnerOrAdmin": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "description": "Login by email, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/roles/{role}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "obtiene un rol",
                "description": "devuelve un rol",
                "operationId": "getRole",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "id del rol a mostrar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CommentRequest": {
                "title": "Store Comment Request",
                "description": "Store Comment request body data",
                "required": [
                    "body,article_id"
                ],
                "properties": {
                    "body": {
                        "title": "body",
                        "description": "cuerpo del comentario",
                        "type": "string",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
                    },
                    "article_id": {
                        "title": "article_id",
                        "description": "id de la noticia a la que pertenece",
                        "type": "string",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CommentRequest"
                }
            },
...
            "CommentResource": {
                "title": "CommentResource",
                "description": "Comment resource",
                "properties": {
                    "data": {
                        "title": "data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Comment"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CommentResource"
                }
            },
            "RoleResource": {
                "title": "RoleResource",
                "description": "Role resource",
                "properties": {
                    "data": {
                        "title": "data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "RoleResource"
                }
            },
...
            "Comment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "body": {
                        "description": "Contenido del comentario",
                        "type": "string",
                        "example": "Loren ipsum..."
                    },
                    "user_status": {
                        "description": "Estado del usuario",
                        "type": "string",
                        "example": "deleted"
                    },
                    "user": {
                        "description": "Usuario que ha creado el comentario",
                        "type": "string",
                        "example": "admin"
                    },
                    "article": {
                        "description": "Noticia a la que pertenece el comentario",
                        "type": "string",
                        "example": ""
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "example": "23-03-09 18:00:00"
                    },
                    "updated_at": {
                        "description": "Fecha de modificación",
                        "type": "string",
                        "example": "23-03-09 18:30:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Comment"
                }
            },
            "Role": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "description": "Nombre del rol",
                        "type": "string",
                        "example": "admin"
                    },
                    "users": {
                        "description": "Array de users con el rol",
                        "type": "string",
                        "example": ""
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "example": "23-03-09 18:00:00"
                    },
                    "updated_at": {
                        "description": "Fecha de modificación",
                        "type": "string",
                        "example": "23-03-09 18:30:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Role"
                }
            },
...
        },
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}